Added solution for Day 15 part 1:
Took a while to realize what the true maximum width was. Code was
working fine otherwise.

+++ AdventOfCode/src/aoc2022/Beacon.java
@@ -2,27 +2,27 @@ package aoc2022;
 
 public class Beacon {
 	
-	private int x;
-	private int y;
+	private long x;
+	private long y;
 	
-	public Beacon(int x, int y) {
+	public Beacon(long x, long y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public long getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(long x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public long getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(long y) {
 		this.y = y;
 	}
